pyproject.toml
[project]
name = "wnucheck"
version = "1.0.0"
description = "WnuCheck app"
authors = []
dependencies = [
    "fastapi>=0.115.6",
    "uvicorn>=0.34.0",
    "colorlog>=6.9.0",
    "python-dotenv>=1.0.1",
    "pytest>=8.3",
    "pytest-cov>=6.0",
    "langchain_core>=0.3.29",
    "python_multipart>=0.0.20",
    "librosa>=0.10.2",
    "openai>=1.59.6",
    "accelerate>=1.2.1",
    "torch>=2.5.1",
    "transformers>=4.48.0"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["backend"]

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = ["tests/backend"]